pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "job-hunt-agent"
version = "0.1.0"
description = "Multi-agent job hunting automation system"
requires-python = ">=3.9"
dependencies = [
    "playwright>=1.40.0",
    "litellm>=1.0.0",
    "gspread>=6.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "chromadb>=0.4.0",
    "python-telegram-bot>=20.0",
    "pyyaml>=6.0",
    "structlog>=24.0.0",
    "rich>=13.0.0",
    "pdfplumber>=0.10.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
